refactor(seed): move seed entry point to cli/ to eliminate src/→test/ import
src/database/seed.ts was the sole cause of src/ importing test/ modules
(seedDatabase from test/seed, truncateAllTables from test/setup). Rather
than moving the seeders into src/, the entry point itself is moved to a
new
cli/ directory since it is dev tooling, not application code — it is
never
invoked by the production runtime (pm2.json only runs index.js and
cron.js).

- Move src/database/seed.ts → cli/seed.ts; update all imports to use
  ../src/ and ../test/ paths
- Move seedMemberUsers() from src/database/seed.ts →
test/seed/user-seeder.ts
  (it is seeder logic and belongs with the other seeders)
- Update test/seed/all.ts to import seedMemberUsers from ./user-seeder
- Update package.json: seed script now runs ts-node cli/seed.ts
- Exclude cli/ from the Docker build context via .dockerignore

Closes #786
