+++ package.json
@@ -9,7 +9,7 @@
     "cron": "nodemon --watch \"src/**\" --ext \"ts,json\" --exec \"ts-node src/cron.ts\"",
     "serve": "node out/src/index.js",
     "schema": "ts-node src/database/schema.ts",
-    "seed": "ts-node src/database/seed.ts",
+    "seed": "ts-node cli/seed.ts",
     "swagger": "ts-node src/start/swagger.ts",
     "migrate": "ts-node src/database/migrate.ts",
     "swagger:validate": "ts-node src/start/swagger.ts && swagger-cli validate ./out/swagger.json",
